this is going to break the database
Added an in-between model called CategoryStrength that serves as way to
hold a list of tuples per word. Each tuple is a 2-tuple that holds a
category and a strong indicator for that category. This should avoid
unnecessary duplication on our end. Each word is will no be mapped to a
list of CategoryWeights (or one), instead of a category.