added subcategories in the form of parent field
Subcategories will be very helpful for avoiding duplication within the
databases. Since there will be a table for each category, words such as
sexual orientation slurs would be duplicated in multiple tables, when
they don't need to be. A strong sexual slur word would always be a
strong slur word. However, the change is to a many-to-many relationship,
so it would follow that a strong sexual slur word would always be a
strong sensual word if that was indicated as a parent. This may not be
desired.